Fixes:    compile with qreal as float Task:     140480
[git-p4: depot-paths = "//depot/qt/main/": change = 244410]

+++ src/svg/qgraphicssvgitem.cpp
@@ -28,7 +28,7 @@ public:
     Q_DECLARE_PUBLIC(QGraphicsSvgItem)
 
     QGraphicsSvgItemPrivate()
-        : renderer(0), maximumCacheSize(1024, 768), shared(false), 
+        : renderer(0), maximumCacheSize(1024, 768), shared(false),
           dirty(true), cached(true)
     {
     }
@@ -166,16 +166,16 @@ static void qt_graphicsItem_highlightSelected(
 {
     const QRectF rect1 = painter->transform().mapRect(QRectF(0, 0, 1, 1));
     const qreal divisor = qMax(rect1.width(), rect1.height());
-    if (qFuzzyCompare(divisor, 0.0))
+    if (qFuzzyCompare(divisor, qreal(0.0)))
         return;
-    const qreal penWidth = 1.0 / divisor;
-    const qreal padFract = 0.05;
+    const qreal penWidth = qreal(1.0) / divisor;
+    const qreal padFract = qreal(0.05);
     const qreal pad =
         qMin(padFract * item->boundingRect().width(), padFract * item->boundingRect().height());
 
     const QRectF rect2 = painter->transform().mapRect(item->boundingRect());
     const qreal minExt = qMin(rect2.width(), rect2.height());
-    if (minExt < 1.0)
+    if (minExt < qreal(1.0))
         return;
 
     painter->setPen(QPen(option->palette.windowText(), penWidth, Qt::DashLine));
@@ -198,9 +198,9 @@ void QGraphicsSvgItem::paint(QPainter *painter, const QStyleOptionGraphicsItem *
 
     QMatrix m = painter->worldMatrix();
     QRect deviceRect = m.mapRect(d->boundingRect).toRect();
-    
+
     if (!d->cached ||
-        deviceRect.size().width() > d->maximumCacheSize.width() || 
+        deviceRect.size().width() > d->maximumCacheSize.width() ||
         deviceRect.size().height() > d->maximumCacheSize.height()) {
         if (d->elemId.isEmpty())
             d->renderer->render(painter, d->boundingRect);
